requirements.txt
Flask
Flask-MySQLdb
qrcode
